+++ packages/core/src/matterbridgeEndpointCommandHandler.ts
@@ -544,6 +544,8 @@ export type CommandHandlerDataMap = {
   'setpointRaiseLower': CommandHandlerData<'Thermostat.setpointRaiseLower'>;
   'setActivePresetRequest': CommandHandlerData<'Thermostat.setActivePresetRequest'>;
   'setActiveScheduleRequest': CommandHandlerData<'Thermostat.setActiveScheduleRequest'>;
+  'addThermostatSuggestion': CommandHandlerData<'Thermostat.addThermostatSuggestion'>;
+  'removeThermostatSuggestion': CommandHandlerData<'Thermostat.removeThermostatSuggestion'>;
   'Thermostat.setpointRaiseLower': {
     command: 'setpointRaiseLower';
     request: Thermostat.SetpointRaiseLowerRequest;
@@ -561,6 +563,20 @@ export type CommandHandlerDataMap = {
   'Thermostat.setActiveScheduleRequest': {
     command: 'setActiveScheduleRequest';
     request: Thermostat.SetActiveScheduleRequest;
+    cluster: 'thermostat';
+    attributes: ClusterAttributeValues<(typeof Thermostat)['attributes']>;
+    endpoint: MatterbridgeEndpoint;
+  };
+  'Thermostat.addThermostatSuggestion': {
+    command: 'addThermostatSuggestion';
+    request: Thermostat.AddThermostatSuggestionRequest;
+    cluster: 'thermostat';
+    attributes: ClusterAttributeValues<(typeof Thermostat)['attributes']>;
+    endpoint: MatterbridgeEndpoint;
+  };
+  'Thermostat.removeThermostatSuggestion': {
+    command: 'removeThermostatSuggestion';
+    request: Thermostat.RemoveThermostatSuggestionRequest;
     cluster: 'thermostat';
     attributes: ClusterAttributeValues<(typeof Thermostat)['attributes']>;
     endpoint: MatterbridgeEndpoint;